# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Margaz Tank Telemetry API
  version: 1.0.0
  description: "API documentation for the Margaz Tank Telemetry backend. Frontend: https://margaz.netlify.app"
  license:
    name: Proprietary
    url: https://margaz.netlify.app
servers:
  - url: http://63.181.47.189
    description: Lightsail direct
...
security:
  - bearerAuth: []
paths:
...
  /api/devices:
    get:
      tags: [Devices]
      summary: List devices
      operationId: listDevices
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Device list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
              example:
                - id: "5ce9d7a1-87b1-4c8d-89b2-3d1b6f60a2d2"
                  deviceId: "1-aktup"
                  name: "Arduino 1-aktup"
                  description: "Otomatik kaydedildi"
                  status: "active"
                  lastSeen: "2026-01-30T18:12:10.501Z"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      tags: [Devices]
      summary: Create device (Admin only)
      operationId: createDevice
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreateRequest'
            example:
              deviceId: "3-aktup"
              name: "Arduino 3-aktup"
              description: "Manuel ekleme"
      responses:
        '201':
          description: Device created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/devices/{id}:
    get:
      tags: [Devices]
      summary: Get device by id
      operationId: getDevice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
              example:
                id: "5ce9d7a1-87b1-4c8d-89b2-3d1b6f60a2d2"
                deviceId: "1-aktup"
                name: "Arduino 1-aktup"
                description: "Otomatik kaydedildi"
                status: "active"
                lastSeen: "2026-01-30T18:12:10.501Z"
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      tags: [Devices]
      summary: Update device (Admin only)
      operationId: updateDevice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceUpdateRequest'
            example:
              name: "Arduino 1-aktup"
              status: "active"
      responses:
        '200':
          description: Device updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Devices]
      summary: Delete device (Admin only)
      operationId: deleteDevice
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/DeviceId'
      responses:
        '200':
          description: Device deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Admin access required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Device not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    DeviceId:
      name: id
      in: path
      required: true
      schema:
        type: string
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Device:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string
        lastSeen:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, deviceId, name, status]
      example:
        id: "5ce9d7a1-87b1-4c8d-89b2-3d1b6f60a2d2"
        deviceId: "1-aktup"
        name: "Arduino 1-aktup"
        description: "Otomatik kaydedildi"
        status: "active"
        lastSeen: "2026-01-30T18:12:10.501Z"
    DeviceCreateRequest:
      type: object
      properties:
        deviceId:
          type: string
        name:
          type: string
        description:
          type: string
      required: [deviceId, name]
    DeviceUpdateRequest:
      type: object
      properties:
        deviceId:
          type: string
        name:
          type: string
        description:
          type: string
        status:
          type: string